+++ src/FunFair.BuildCheck.ProjectChecks/Settings/PublishableProjectsThatPassTrimAnalyzerShouldHavePublishedTrimmedSetPolicy.cs
@@ -1,0 +1,26 @@
+using System;
+using System.Xml;
+using FunFair.BuildCheck.ProjectChecks.Helpers;
+using Microsoft.Extensions.Logging;
+
+namespace FunFair.BuildCheck.ProjectChecks.Settings;
+
+public sealed class PublishableProjectsThatPassTrimAnalyzerShouldHavePublishedTrimmedSetPolicy : SimplePropertyProjectCheckBase
+{
+    public PublishableProjectsThatPassTrimAnalyzerShouldHavePublishedTrimmedSetPolicy(ILogger<PublishableProjectsThatPassTrimAnalyzerShouldHavePublishedTrimmedSetPolicy> logger)
+        : base(propertyName: "PublishTrimmed", requiredValue: "true", logger: logger)
+    {
+    }
+
+    protected override bool CanCheck(string projectName, string projectFolder, XmlDocument project)
+    {
+        if (!project.IsPublishable())
+        {
+            return false;
+        }
+
+        string? value = project.GetProperty("EnableTrimAnalyzer");
+
+        return !string.IsNullOrWhiteSpace(value) && StringComparer.InvariantCultureIgnoreCase.Equals(x: value, y: "true");
+    }
+}